servegcs: fix handling of Range-free requests
Forcing the range in the blobstore header was necessary
to make If-Range work properly, but doing it when Range was
not in the request was making blobstore respond with 206
instead of 200.

This confused some RSS feed readers doing a plain
GET https://research.swtch.com/feed.atom
and expecting a 200 response.

Thanks to Ralph Corderoy for noticing.