servegcs: stop using Blobstore APIs
The original AppEngine implementation provided a nice hack that if you
set an X-AppEngine-BlobKey header in the response, the server itself
would take care of redirecting the request to the Blobstore implementation
for handling, all server side, without the client seeing any redirects.
Later, Blobstore let you create a key for a Google Cloud Storage object,
so you could use this trick to serve GCS objects "out of server".

That server is gone, and Blobstore might as well be too.
Rewrite the code to forward the request to GCS and explicitly
ferry bytes back and forth, like the old server used to do for us.